deleting the emaial mutator and overriding the "save" method on the model to achieve the same fuctionality. and fixing a bug with email mutator.

+++ src/Pingpong/Confirmation/Confirmable.php
@@ -4,6 +4,22 @@ namespace Pingpong\Confirmation;
 
 trait Confirmable
 {
+    public function save(array $options)
+    {
+        // We throw any new email address in the unconfirmed_emails table.
+        $confirmation = [
+            'user_id' => $this->id,
+            'confirmation_token' => str_random(16),
+            'email' => $this->attributes['email']
+        ];
+
+        \DB::table('unconfirmed_emails')->insert($confirmation);
+        $this->attributes['email'] = null;
+
+        parent::save($options);
+
+    }
+
     /**
      * Accessor for 'confirmation_url' attribute.
      *
@@ -30,23 +46,6 @@ trait Confirmable
 
     }
 
-    /**
-     * Mutator for 'email' attribute.
-     *
-     * @return string
-     */
-    public function setEmailAttribute($email)
-    {
-        // We throw any new email address in the unconfirmed_emails table.
-        $confirmation = [
-            'user_id' => $this->id,
-            'confirmation_token' => str_random(16),
-            'email' => $email
-        ];
-
-        \DB::table('unconfirmed_emails')->insert($confirmation);
-        $this->attributes['email'] = null;
-    }
 
     /**
      * Confirm current user.
@@ -56,12 +55,16 @@ trait Confirmable
      */
     public function confirm($token)
     {
+        // Now we transfer the email value to the actual table. (unconfirmed_emails => )
         $email = \DB::table('unconfirmed_emails')->where('confirmation_token', $token)->value('email');
-        \DB::table('unconfirmed_emails')->where('confirmation_token', $token)->delete();
 
-        // This way we can bypass the "email" Mutator.
-        // And inject the confirmed email address in users table.
-        return \DB::table($this->getTable())->where('id', $this->id)->update(['email' => $email]);
+        if ($email) {
+
+            \DB::table('unconfirmed_emails')->where('confirmation_token', $token)->delete();
+
+            // inject the confirmed email address in users table.
+            return \DB::table($this->getTable())->where('id', $this->id)->update(['email' => $email]);
+        }
     }
